set up deposit dialog fragment
