+++ app/src/main/java/com/ejsfbu/app_main/Activities/ParentProfileActivity.java
@@ -15,11 +15,11 @@ import com.bumptech.glide.load.engine.DiskCacheStrategy;
 import com.bumptech.glide.load.resource.bitmap.CenterCrop;
 import com.bumptech.glide.load.resource.bitmap.CircleCrop;
 import com.bumptech.glide.request.RequestOptions;
-import com.ejsfbu.app_main.EditFragments.EditEmailDialogFragment;
-import com.ejsfbu.app_main.EditFragments.EditNameDialogFragment;
-import com.ejsfbu.app_main.EditFragments.EditPasswordDialogFragment;
-import com.ejsfbu.app_main.EditFragments.EditProfileImageDialogFragment;
-import com.ejsfbu.app_main.EditFragments.EditUsernameDialogFragment;
+import com.ejsfbu.app_main.DialogFragments.EditEmailDialogFragment;
+import com.ejsfbu.app_main.DialogFragments.EditNameDialogFragment;
+import com.ejsfbu.app_main.DialogFragments.EditPasswordDialogFragment;
+import com.ejsfbu.app_main.DialogFragments.EditProfileImageDialogFragment;
+import com.ejsfbu.app_main.DialogFragments.EditUsernameDialogFragment;
 import com.ejsfbu.app_main.R;
 import com.ejsfbu.app_main.models.User;
 import com.parse.ParseFile;
