+++ app/src/main/java/com/ejsfbu/app_main/DialogFragments/EditProfileImageDialogFragment.java
@@ -1,4 +1,4 @@
-package com.ejsfbu.app_main.EditFragments;
+package com.ejsfbu.app_main.DialogFragments;
 
 import android.Manifest;
 import android.app.Activity;
