+++ app/src/main/java/com/ejsfbu/app_main/DialogFragments/RemoveBankDialogFragment.java
@@ -1,0 +1,102 @@
+package com.ejsfbu.app_main.DialogFragments;
+
+import android.content.Context;
+import android.graphics.Point;
+import android.os.Bundle;
+import android.view.Display;
+import android.view.Gravity;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.view.Window;
+import android.view.WindowManager;
+import android.widget.Button;
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.DialogFragment;
+
+import com.ejsfbu.app_main.R;
+import com.ejsfbu.app_main.models.User;
+import com.parse.ParseUser;
+
+public class RemoveBankDialogFragment extends DialogFragment {
+    // View objects
+    private Button bConfirm;
+    private Button bCancel;
+    private Context context;
+    private User user;
+
+    public RemoveBankDialogFragment() {
+        // Empty constructor is required for DialogFragment
+        // Make sure not to add arguments to the constructor
+        // Use `newInstance` instead as shown below
+    }
+
+    public static RemoveBankDialogFragment newInstance(String title) {
+        RemoveBankDialogFragment frag = new RemoveBankDialogFragment();
+        Bundle args = new Bundle();
+        args.putString("title", title);
+        frag.setArguments(args);
+        return frag;
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        context = getContext();
+        return inflater.inflate(R.layout.fragment_edit_bank, container);
+    }
+
+    @Override
+    public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+        user = (User) ParseUser.getCurrentUser();
+        // Get field from view
+        bConfirm = view.findViewById(R.id.bConfirm);
+        bCancel = view.findViewById(R.id.bCancel);
+        // Fetch arguments from bundle and set title
+        String title = getArguments().getString("title", "Enter Name");
+        getDialog().setTitle(title);
+        // Show soft keyboard automatically and request focus to field
+        getDialog().getWindow().setSoftInputMode(
+                WindowManager.LayoutParams.SOFT_INPUT_STATE_VISIBLE);
+        setOnClick();
+    }
+
+    // Defines the listener interface
+    public interface RemoveBankDialogListener {
+        void onFinishEditDialog();
+    }
+
+    // Call this method to send the data back to the parent fragment
+    public void sendBackResult() {
+        // Notice the use of `getTargetFragment` which will be set when the dialog is displayed
+        RemoveBankDialogListener listener = (RemoveBankDialogListener) getFragmentManager().findFragmentById(R.id.flContainer);
+        listener.onFinishEditDialog();
+        dismiss();
+    }
+
+
+    public void onResume() {
+        // Store access variables for window and blank point
+        Window window = getDialog().getWindow();
+        Point size = new Point();
+        // Store dimensions of the screen in `size`
+        Display display = window.getWindowManager().getDefaultDisplay();
+        display.getSize(size);
+        window.setLayout((int) (size.x * 0.85), WindowManager.LayoutParams.WRAP_CONTENT);
+        window.setGravity(Gravity.CENTER);
+        // Call super onResume after sizing
+        super.onResume();
+    }
+
+    private void setOnClick() {
+        bCancel.setOnClickListener(view -> {
+            dismiss();
+        });
+
+        bConfirm.setOnClickListener(view -> {
+            sendBackResult();
+        });
+    }
+}
