+++ app/src/main/java/com/ejsfbu/app_main/Fragments/GoalDetailsFragment.java
@@ -1,12 +1,9 @@
 package com.ejsfbu.app_main.Fragments;
 
-import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Adapter;
 import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.ProgressBar;
@@ -16,22 +13,20 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.load.engine.DiskCacheStrategy;
 import com.bumptech.glide.load.resource.bitmap.CenterCrop;
 import com.bumptech.glide.load.resource.bitmap.CircleCrop;
 import com.bumptech.glide.request.RequestOptions;
-import com.ejsfbu.app_main.Activities.AddGoalActivity;
+import com.ejsfbu.app_main.Activities.MainActivity;
 import com.ejsfbu.app_main.Adapters.GoalAdapter;
 import com.ejsfbu.app_main.Adapters.TransactionAdapter;
+import com.ejsfbu.app_main.DialogFragments.RemoveBankDialogFragment;
 import com.ejsfbu.app_main.EndlessRecyclerViewScrollListener;
 import com.ejsfbu.app_main.R;
 import com.ejsfbu.app_main.models.Goal;
 import com.ejsfbu.app_main.models.Transaction;
-import com.parse.FindCallback;
-import com.parse.ParseException;
 import com.parse.ParseFile;
 
 import java.util.ArrayList;
@@ -39,6 +34,7 @@ import java.util.List;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
+import butterknife.OnClick;
 import butterknife.Unbinder;
 
 import java.text.NumberFormat;
@@ -75,40 +71,7 @@ public class GoalDetailsFragment extends Fragment {
         unbinder = ButterKnife.bind(this, view);
 
         Goal goal = getArguments().getParcelable("Goal");
-
-        ParseFile image = goal.getParseFile("image");
-        if (image != null) {
-            String imageUrl = image.getUrl();
-            imageUrl = imageUrl.substring(4);
-            imageUrl = "https" + imageUrl;
-            RequestOptions options = new RequestOptions();
-            options.placeholder(R.drawable.ic_iconfinder_icons_user_1564534)
-                    .diskCacheStrategy(DiskCacheStrategy.AUTOMATIC)
-                    .error(R.drawable.ic_iconfinder_icons_user_1564534)
-                    .transform(new CenterCrop())
-                    .transform(new CircleCrop());
-            Glide.with(getContext())
-                    .load(imageUrl)
-                    .apply(options) // Extra: round image corners
-                    .into(ivGoalDetailsImage);
-        }
-
         setGoalInfo(goal);
-
-
-        tvGoalDetailsName.setText(goal.getName());
-        //TODO Format the Date
-        tvTotalCost.setText("$" + goal.getCost());
-        String endDate = goal.get("endDate").toString();
-        String finalizedDate = endDate.substring(4,10) + ", " + endDate.substring(24,28);
-        System.out.println(endDate.length());
-        System.out.println(endDate);
-        tvCompletionDate.setText(finalizedDate);
-
-        Double percentDone = (goal.getSaved() / goal.getCost()) * 100;
-        tvDetailsPercentDone.setText(String.format("%.1f", percentDone.floatValue()) + "%");
-        pbDetailsPercentDone.setProgress((int) percentDone.doubleValue());
-        pbDetailsPercentDone.getProgressDrawable().setTint(getContext().getResources().getColor(R.color.money_green));
 
         transactionsList = new ArrayList<>();
         adapter = new TransactionAdapter(getContext(), transactionsList);
@@ -142,6 +105,23 @@ public class GoalDetailsFragment extends Fragment {
         tvDetailsPercentDone.setText(String.format("%.1f", percentDone.floatValue()) + "%");
         pbDetailsPercentDone.setProgress((int) percentDone.doubleValue());
         pbDetailsPercentDone.getProgressDrawable().setTint(getContext().getResources().getColor(R.color.money_green));
+
+        ParseFile image = goal.getParseFile("image");
+        if (image != null) {
+            String imageUrl = image.getUrl();
+            imageUrl = imageUrl.substring(4);
+            imageUrl = "https" + imageUrl;
+            RequestOptions options = new RequestOptions();
+            options.placeholder(R.drawable.ic_iconfinder_icons_user_1564534)
+                    .diskCacheStrategy(DiskCacheStrategy.AUTOMATIC)
+                    .error(R.drawable.ic_iconfinder_icons_user_1564534)
+                    .transform(new CenterCrop())
+                    .transform(new CircleCrop());
+            Glide.with(getContext())
+                    .load(imageUrl)
+                    .apply(options) // Extra: round image corners
+                    .into(ivGoalDetailsImage);
+        }
     }
 
     public String formatDate(Goal goal) {
@@ -163,5 +143,18 @@ public class GoalDetailsFragment extends Fragment {
         NumberFormat currency = NumberFormat.getCurrencyInstance(Locale.US);
         String convertedAmount = currency.format(amount);
         return convertedAmount;
+    }
+
+    @OnClick(R.id.deposit_btn)
+    public void onClickDeposit() {
+        showDepositDialog();
+    }
+
+    // Call this method to launch the edit dialog
+    private void showDepositDialog() {
+        RemoveBankDialogFragment removeBankDialogFragment
+                = RemoveBankDialogFragment.newInstance("Remove bank");
+        removeBankDialogFragment.show(MainActivity.fragmentManager,
+                "fragment_edit_bank");
     }
 }
