+++ app/src/main/java/com/ejsfbu/app_main/Fragments/ProfileFragment.java
@@ -25,11 +25,11 @@ import com.bumptech.glide.request.RequestOptions;
 import com.ejsfbu.app_main.Activities.LoginActivity;
 import com.ejsfbu.app_main.Activities.MainActivity;
 import com.ejsfbu.app_main.Adapters.ParentDisplayAdapter;
-import com.ejsfbu.app_main.EditFragments.EditEmailDialogFragment;
-import com.ejsfbu.app_main.EditFragments.EditNameDialogFragment;
-import com.ejsfbu.app_main.EditFragments.EditPasswordDialogFragment;
-import com.ejsfbu.app_main.EditFragments.EditProfileImageDialogFragment;
-import com.ejsfbu.app_main.EditFragments.EditUsernameDialogFragment;
+import com.ejsfbu.app_main.DialogFragments.EditEmailDialogFragment;
+import com.ejsfbu.app_main.DialogFragments.EditNameDialogFragment;
+import com.ejsfbu.app_main.DialogFragments.EditPasswordDialogFragment;
+import com.ejsfbu.app_main.DialogFragments.EditProfileImageDialogFragment;
+import com.ejsfbu.app_main.DialogFragments.EditUsernameDialogFragment;
 import com.ejsfbu.app_main.R;
 import com.ejsfbu.app_main.models.User;
 import com.parse.FindCallback;
